renamed the homepage.jsx
