+++ src/App.jsx
@@ -3,8 +3,9 @@ import React from 'react';
 import { BrowserRouter as Router, Route, Switch, Routes } from 'react-router-dom';
 import header from '../src/Hompage/header';
 import Navbar from '../src/Hompage/navbar';
-import Home from './pages/hompage';
+import Home from './pages/homepage';
 import Login from './pages/Profile/login';
+import footer from '../src/Hompage/footer';
 
 function App() {
   return (
@@ -19,6 +20,9 @@ function App() {
             <Route path="/about" element={<AboutUs/>} />
             <Route path="/page/profile" element={<Profile/>} />
             <Route path="/page/Exercises" element={<Excercises/>} />
+            <Route path="/page/Foods" element={<Foods/>} />
+
+            
           </Routes>
         </Wrapper>
         <Footer />
